+++ data/colors_v2/vars/component_dark.ts
@@ -8,15 +8,26 @@ export default {
     stackFadeMore: get('scale.gray.7'),
     childShadow: (theme: any) => `-2px -2px 0 ${get('scale.gray.9')(theme)}`
   },
+  diffstat: {
+    deletionBorder: get('border.subtle'),
+    additionBorder: get('border.subtle'),
+  },
   selectMenu: {
     backdropBorder: get('scale.gray.5'),
     tapHighlight: alpha(get('scale.gray.6'), 0.5),
     tapFocusBg: get('scale.blue.8')
   },
+  overlay: {
+    shadow: (theme: any) => `0 0 0 1px ${get('scale.gray.6')(theme)}, 0 16px 32px ${alpha(get('scale.black'), 0.85)(theme)}`,
+  },
   header: {
     text: alpha(get('scale.white'), 0.7),
     bg: get('scale.gray.8'),
-    logo: get('scale.gray.0')
+    logo: get('scale.gray.0'),
+    search: {
+      bg: get('scale.gray.9'),
+      border: get('scale.gray.6')
+    }
   },
   sidenav: {
     selectedBg: get('scale.gray.7')
@@ -103,24 +114,24 @@ export default {
 
     danger: {
       text: get('scale.red.4'),
-      hoverText: '#ffffff',
+      hoverText: get('scale.gray.9'),
       hoverBg: get('scale.red.5'),
-      hoverBorder: get('border.subtle'),
+      hoverBorder: get('scale.red.2'),
       hoverShadow: '0 0 transparent',
       hoverInsetShadow: '0 0 transparent',
+      hoverIcon: get('scale.gray.9'),
       hoverCounterBg: alpha('#fff', 0.2),
       selectedText: '#ffffff',
       selectedBg: get('scale.red.6'),
-      selectedBorder: get('border.subtle'),
+      selectedBorder: get('scale.red.3'),
       selectedShadow: '0 0 transparent',
       disabledText: alpha(get('scale.red.4'), 0.5),
       disabledBg: get('scale.gray.9'),
       disabledCounterBg: alpha(get('scale.red.5'), 0.05),
-      focusBorder: get('border.subtle'),
-      focusShadow: (theme: any) => `0 0 0 3px ${alpha(get('scale.red.6'), 0.4)(theme)}`,
+      focusBorder: get('scale.red.4'),
+      focusShadow: (theme: any) => `0 0 0 3px ${alpha(get('scale.red.4'), 0.4)(theme)}`,
       counterBg: alpha(get('scale.red.5'), 0.1),
-      icon: get('scale.red.4'),
-      hoverIcon: get('scale.white')
+      icon: get('scale.red.4')
     }
   }
 }