+++ data/colors_v2/vars/component_light.ts
@@ -8,15 +8,26 @@ export default {
     stackFadeMore: get('scale.gray.2'),
     childShadow: (theme: any) => `-2px -2px 0 ${alpha(get('scale.white'), 0.8)(theme)}`
   },
+  diffstat: {
+    deletionBorder: get('border.subtle'),
+    additionBorder: get('border.subtle'),
+  },
   selectMenu: {
     backdropBorder: 'transparent',
     tapHighlight: alpha(get('scale.gray.3'), 0.5),
     tapFocusBg: get('scale.blue.1')
   },
+  overlay: {
+    shadow: (theme: any) => `0 1px 3px ${alpha(get('scale.black'), 0.12)(theme)}, 0 8px 24px ${alpha(get('scale.gray.7'), 0.12)(theme)}`,
+  },
   header: {
     text: alpha(get('scale.white'), 0.7),
     bg: get('scale.gray.9'),
-    logo: get('scale.white')
+    logo: get('scale.white'),
+    search: {
+      bg: get('scale.gray.9'),
+      border: get('scale.gray.6')
+    }
   },
   sidenav: {
     selectedBg: get('scale.white')
@@ -65,7 +76,7 @@ export default {
 
     primary: {
       text: get('scale.white'),
-      bg: '#2ea44f',
+      bg: get('scale.green.4'),
       border: get('border.subtle'),
       shadow: (theme: any) => `0 1px 0 ${alpha(get('scale.black'), 0.1)(theme)}`,
       insetShadow: (theme: any) => `inset 0 1px 0 ${alpha(get('scale.white'), 0.03)(theme)}`,
@@ -76,7 +87,7 @@ export default {
       disabledText: alpha(get('scale.white'), 0.8),
       disabledBg: '#94d3a2',
       disabledBorder: get('border.subtle'),
-      focusBg: '#2ea44f',
+      focusBg: get('scale.green.4'),
       focusBorder: get('border.subtle'),
       focusShadow: (theme: any) => `0 0 0 3px ${alpha(get('btn.primary.focusBg'), 0.4)(theme)}`,
       icon: alpha(get('scale.white'), 0.8),