+++ data/colors_v2/vars/global_dark.ts
@@ -28,7 +28,7 @@ export default {
   // Roles
   neutral: {
     emphasisPlus: get('scale.gray.4'),
-    emphasis: get('scale.gray.5'),
+    emphasis: get('scale.gray.4'),
     muted: alpha(get('scale.gray.4'), 0.4),
     subtle: alpha(get('scale.gray.4'), 0.1)
   },
@@ -82,7 +82,7 @@ export default {
       sticky: alpha(get('scale.gray.9'), 0.95) // use for sticky headers
     },
     border: {
-      active: '#F78166', // coral
+      active: get('scale.coral.3'),
       contrast: alpha(get('scale.white'), 0.2) // use to increase contrast
     },
     shadow: {